Improved warp distinction

+++ src/effect-GL.c
@@ -28,11 +28,11 @@ void pok_fadeout_effect_render(struct pok_graphics_subsystem* sys,const struct p
             { /* 10 verteces define 4 quadrilaterals */
                 glVertex2f(0.0,0.0);
                 glVertex2f(effect->hs[2],effect->hs[0]);
-                glVertex2f(sys->wwidth,0.0);
+                glVertex2f((GLfloat)sys->wwidth,0.0);
                 glVertex2f(effect->hs[3],effect->hs[0]);
-                glVertex2f(sys->wwidth,sys->wheight);
+                glVertex2f((GLfloat)sys->wwidth,sys->wheight);
                 glVertex2f(effect->hs[3],effect->hs[1]);
-                glVertex2f(0.0,sys->wheight);
+                glVertex2f(0.0,(GLfloat)sys->wheight);
                 glVertex2f(effect->hs[2],effect->hs[1]);
                 glVertex2f(0.0,0.0);
                 glVertex2f(effect->hs[2],effect->hs[0]);
